Split config into fallback- and user-config
The resources files are in general not writable. So implement an override
mechanism, where changes in a user config file overrides changes in the
fallback config.

I tried to be thorough with determining the user config path. Better get
this right the first time than changing the path again later and breaking
all user configs.
The MacOS config path detection is still to be done.